Generate bytecode for storing and loadin of variables
Use getFreshVar to get a fresh int for slot number. On assign,
let generateExprCode ensure that the value to be stored is at
the top of the stack. Then use IStore/AStore to store the top
of the stack to the slot number stored for that VariableSymbol.

A new Map[VariableSymbol, Int] is introduced to keep track of
store numbers of VariableSymbols.