Use overridden field of MethodSymbols
Extremely ugly implementation that will make the DrawStuff program run. Simply works by checking through the methods of the currentclass to see if there is an overriding method, if there is, check the method's symbol's classSymbol's name and use it as object type of the method call. currentClass will be the class invoking the method but if there is a self reference in the method, it will have the type of the superclass. Fck that.